Fix windowless context draw issue on OSX
Lack of double buffering was the reason why our rendering was failing.

According to the Apple documentation, the [NSOpenGLContext flushBuffer]
will do nothing if double/triple buffering is not enabled, therefore it
won't paint to the window.  Furthermore, the reason why it was painting
to only one window rather was because apparently glFlush is actually
what is used on OSX to draw to a window (which is strange because that
function is not supposed to be used to do that).  The gs_flush command
in obs-video.c was causing the last obs_display that rendered to
display.