libobs-opengl: Don't call glFlush w/no swap on OSX
This function can have different implementations depending on the
operating system, and on OSX, if there's a view assigned to the context,
it will additionally be used to draw that view.  If there is no view
assigned to that context then it will actually be used to flush
commands.  Any extra flushes that we do with a view enabled on OSX are
superfluous and can potentially negatively impact performance.